Modularizing the .py files to work accross LSTM methods
